+++ R/seq_linter.R
@@ -1,6 +1,7 @@
 #' Sequence linter
 #'
-#' Check for `1:length(...)`, `1:nrow(...)`, `1:ncol(...)`, `1:NROW(...)` and `1:NCOL(...)` expressions.
+#' Check for `1:length(...)`, `1:nrow(...)`, `1:ncol(...)`, `1:NROW(...)` and `1:NCOL(...)` expressions in base-R.
+#' Additionally check `1:n()` (from dplyr) and `1:.N` (from data.table).
 #' These often cause bugs when the right-hand side is zero.
 #' It is safer to use [base::seq_len()] or [base::seq_along()] instead.
 #'
@@ -8,18 +9,30 @@
 #' @seealso [linters] for a complete list of linters available in lintr.
 #' @export
 seq_linter <- function() {
-  bad_funcs <- c("length", "nrow", "ncol", "NROW", "NCOL", "dim")
+  bad_funcs <- c("length", "n", "nrow", "ncol", "NROW", "NCOL", "dim")
 
+  # `.N` from {data.table} is special since it's not a function but a symbol
   xpath <- glue::glue("//expr[
     expr[NUM_CONST[text() =  '1' or text() =  '1L']]
     and OP-COLON
-    and expr[expr[(expr|self::*)[SYMBOL_FUNCTION_CALL[ {xp_text_in_table(bad_funcs)} ]]]]
+    and (
+          expr[expr[(expr|self::*)[SYMBOL_FUNCTION_CALL[ {xp_text_in_table(bad_funcs)} ]]]]
+          or
+          expr[SYMBOL = '.N']
+        )
   ]")
 
   ## The actual order of the nodes is document order
   ## In practice we need to handle length(x):1
   get_fun <- function(expr, n) {
     funcall <- xml2::xml_find_chr(expr, sprintf("string(./expr[%d])", n))
+
+    # `dplyr::n()` is special because it has no arguments, so the lint message
+    # should mention `n()`, and not `n(...)`
+    if (identical(funcall, "n()")) {
+      return(funcall)
+    }
+
     fun <- gsub("\\(.*\\)", "(...)", trimws(funcall))
     bad_fun <- fun %in% bad_funcs
     fun[bad_fun] <- paste0(fun[bad_fun], "(...)")